Split video_driver into one file per class
