+++ game/dancegraph.cc
@@ -1,6 +1,7 @@
 #include "dancegraph.hh"
 #include "song.hh"
 #include "i18n.hh"
+#include "graphic/view_trans.hh"
 
 #include <stdexcept>
 #include <algorithm>
