+++ game/glutil.cc
@@ -1,5 +1,6 @@
 #include "glutil.hh"
-#include "video_driver.hh"
+#include "graphic/video_driver.hh"
+#include "graphic/window.hh"
 
 namespace glutil {
