+++ game/graphic/lyrics_color_trans.hh
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "glmath.hh"
+
+struct Color;
+class Window;
+
+struct LyricColorTrans {
+	LyricColorTrans(Window&, Color const& fill, Color const& stroke, Color const& newFill, Color const& newStroke);
+	~LyricColorTrans();
+
+  private:
+	Window& m_window;
+	glmath::vec4 oldFill;
+	glmath::vec4 oldStroke;
+};
