+++ game/guitargraph.cc
@@ -2,6 +2,7 @@
 #include "fs.hh"
 #include "song.hh"
 #include "i18n.hh"
+#include "graphic/view_trans.hh"
 
 #include <cmath>
 #include <cstdlib>
@@ -59,10 +60,10 @@ void GuitarGraph::initGuitar() {
 	// Copy all tracks of guitar types (not DRUMS and not KEYBOARD) to m_instrumentTracks
 	for (auto const& elem: m_song.instrumentTracks) {
 		std::string index = elem.first;
-		if (index != TrackName::DRUMS 
-			&& index != TrackName::DRUMS_SNARE 
-			&& index != TrackName::DRUMS_CYMBALS 
-			&& index != TrackName::DRUMS_TOMS 
+		if (index != TrackName::DRUMS
+			&& index != TrackName::DRUMS_SNARE
+			&& index != TrackName::DRUMS_CYMBALS
+			&& index != TrackName::DRUMS_TOMS
 			&& index != TrackName::KEYBOARD)  {
 				m_instrumentTracks[index] = &elem.second;
 			}
@@ -82,9 +83,9 @@ void GuitarGraph::initDrums() {
 	// Copy all tracks of drum type to m_instrumentTracks
 	for (auto const& elem: m_song.instrumentTracks) {
 		std::string index = elem.first;
-		if (index == TrackName::DRUMS 
-			|| index == TrackName::DRUMS_SNARE 
-			|| index == TrackName::DRUMS_CYMBALS 
+		if (index == TrackName::DRUMS
+			|| index == TrackName::DRUMS_SNARE
+			|| index == TrackName::DRUMS_CYMBALS
 			|| index == TrackName::DRUMS_TOMS ) {
 				m_instrumentTracks[index] = &elem.second;
 			}
